fix: make sure we don't modify reports not owned by kyverno
